Fix bugs
Bug 1: after "=" is clicked, result was not used.
Set up a boolean to determine if user is inputting
an entirely new operation or continuing to use the
previous result.

Bug 2: wrong designation of firstVar and secondVar
after decimal point is clicked. Omit an
unnecessary conditional which checks if firstVar
is an empty string.